refactor: update names of const variables and change int to uint16_t
In some places in code int has to big size Better option is uint16_t
Thank's this use of RAM will be better optimalized

+++ src/Message.cpp
@@ -45,7 +45,7 @@ void Message::deserializeMessage(String message, JsonDocument &doc) //deserializ
     }
 }
 
-int Message::getMessageTypeID(JsonDocument &doc) //get ID of message type
+uint16_t Message::getMessageTypeID(JsonDocument &doc) //get ID of message type
 {
     String type = doc[CommSymbol::getSymbol(CommSymbolID::messageType)]; //get type of message
